Refactor recipes from file into submodule
