+++ docs/api/examol.store.rst
@@ -19,3 +19,26 @@ examol.store.recipes
 .. automodule:: examol.store.recipes
    :members:
    :show-inheritance:
+
+examol.store.recipes.base
++++++++++++++++++++++++++
+
+.. automodule:: examol.store.recipes.base
+   :members:
+   :show-inheritance:
+
+
+examol.store.recipes.basic
+++++++++++++++++++++++++++
+
+.. automodule:: examol.store.recipes.basic
+   :members:
+   :show-inheritance:
+
+
+examol.store.recipes.base
++++++++++++++++++++++++++
+
+.. automodule:: examol.store.recipes.basic
+   :members:
+   :show-inheritance:
