+++ examol/select/base.py
@@ -7,7 +7,7 @@ from typing import Iterator, Sequence
 import numpy as np
 
 from examol.store.models import MoleculeRecord
-from examol.store.recipes import PropertyRecipe
+from examol.store.recipes.base import PropertyRecipe
 
 logger = logging.getLogger(__name__)
