+++ examol/select/botorch.py
@@ -19,7 +19,7 @@ import torch
 
 from examol.select.base import RankingSelector, _extract_observations
 from examol.store.models import MoleculeRecord
-from examol.store.recipes import PropertyRecipe
+from examol.store.recipes.base import PropertyRecipe
 
 
 class _EnsembleCovarianceModel(Model):
