+++ examol/store/recipes/__init__.py
@@ -1,0 +1,1 @@
+"""Tools for computing the properties of molecules from their record"""
