+++ examples/redoxmers-bebop/spec.py
@@ -12,7 +12,7 @@ from examol.score.rdkit import RDKitScorer, make_gpr_model
 from examol.simulate.ase import ASESimulator
 from examol.start.fast import RandomStarter
 from examol.steer.single import SingleStepThinker
-from examol.store.recipes import RedoxEnergy
+from examol.store.recipes.redox import RedoxEnergy
 from examol.select.bayes import ExpectedImprovement
 from examol.specify import ExaMolSpecification
