+++ examples/redoxmers/run/report.md
@@ -1,14 +1,12 @@
 # Run Report
-Report time: 2023-10-02 17:57:52.786681
+Report time: 2023-10-06 10:02:29.180054
 
 ## Task Summary
 Measures how many tasks have run as part of the application
 
 | Task Type   |   Count |   Node Hours | Failures   |
 |-------------|---------|--------------|------------|
-| inference   |       5 |      0.00033 | 0 (0.0%)   |
-| train       |       5 |      0.00015 | 0 (0.0%)   |
-| simulation  |      36 |      0.02    | 0 (0.0%)   |
+| simulation  |       2 |      0.00091 | 0 (0.0%)   |
 
 ## Outcomes over Time
 The property of the molecules over time.
