+++ tests/score/conftest.py
@@ -1,7 +1,7 @@
 from pytest import fixture
 
 from examol.store.models import MoleculeRecord
-from examol.store.recipes import PropertyRecipe
+from examol.store.recipes.base import PropertyRecipe
 
 
 class FakeRecipe(PropertyRecipe):
