+++ tests/specify/test_spec.py
@@ -15,7 +15,7 @@ from examol.simulate.ase import ASESimulator
 from examol.specify import ExaMolSpecification
 from examol.steer.single import SingleStepThinker
 from examol.store.models import MoleculeRecord
-from examol.store.recipes import RedoxEnergy
+from examol.store.recipes.redox import RedoxEnergy
 
 
 @fixture()
